fix：サインアップのバグ
