+++ views/login.ejs
@@ -21,7 +21,7 @@
         </form>
         <hr>
         <h1>Sign up</h1>
-        <form action="/signup/" method="post">
+        <form action="auth/signup/" method="post">
             <section>
                 <label for="id">id</label>
                 <input id="id" name="id" type="text" required>
